requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9
Pillow>=10.0
numpy>=1.24,<2.0
bcrypt>=4.1
PyJWT>=2.8
pydantic>=2.6
